+++ python/cuml/cluster/kmeans.pyx
@@ -22,6 +22,7 @@
 import ctypes
 import cudf
 import numpy as np
+import warnings
 
 from numba import cuda
 from cuml import numba_utils
@@ -169,6 +170,10 @@ class KMeans:
         self.cluster_centers_ = None
         self.n_gpu = n_gpu
         self.gpu_id = gpu_id
+        warnings.warn("The current version of the KMeans algorithm will change to be based on cuML's machine"
+                      " learning primitives to achieve performance and reliability improvements."
+                      "Some functionality may be affected.",
+                      FutureWarning)
 
     def _get_ctype_ptr(self, obj):
         # The manner to access the pointers in the gdf's might change, so
@@ -446,7 +451,7 @@ class KMeans:
         variables = [ 'algorithm','copy_x','init','max_iter','n_clusters','n_init','n_jobs','precompute_distances','random_state','tol','verbose']
         for key in variables:
             var_value = getattr(self,key,None)
-            params[key] = var_value   
+            params[key] = var_value
         return params